[refactor] textService, RichMediaService - 기능에 따라 서비스를 분리함
